Extract html fixture renderers
