+++ lib/html_fixtures/renderer.rb
@@ -1,55 +1,32 @@
 module HtmlFixtures
   class Renderer
     def render_all
+      [
+        HtmlFixtures::Renderers::DarkmodeSwitch.new,
+        HtmlFixtures::Renderers::SearchesCombobox.new
+      ].each do |renderer|
+        File.write(renderer.fixture_path, renderer.render(view_context))
+      end
+    end
+
+    def view_context
+      @view_context ||= build_view_context
+    end
+
+    def build_view_context
       request = ActionDispatch::Request.new({})
       request.routes = ApplicationController._routes
 
       instance = ApplicationController.new
       instance.set_request! request
       instance.set_response! ApplicationController.make_response!(request)
-      view_context = instance.view_context
-
-      File.write(
-        Rails.root.join("app", "javascript", "test", "fixtures", "views", "darkmode", "switch.html"),
-        DarkMode::Switch.call
-      )
-
-      search_result = Data.define(:title, :url, :body) do
-        def to_key = [url.parameterize]
-
-        def model_name = ActiveModel::Name.new(self, nil, "SearchResult")
-      end
-
-      Class.new(ApplicationComponent) do
-        attr_reader :search_result
-
-        def initialize(search_result)
-          @search_result = search_result
-        end
-
-        def view_template
-          a(href: search_result.url) do
-            div { raw safe(search_result.title) }
-            div { raw safe(search_result.body) }
-          end
-        end
-
-        Searches::Result.register(search_result, self)
-      end
-
-      search_results = [
-        search_result.new(title: "Here’the thing", url: "/articles/heres-the-thing", body: "Ruby is a programming language"),
-        search_result.new(title: "Introducing Joy of Rails", url: "/articles/joy-of-rails", body: "Rails is a web application framework")
-      ]
-
-      File.write(
-        Rails.root.join("app", "javascript", "test", "fixtures", "views", "searches", "combobox.html"),
-        Searches::Combobox.new(results: search_results, query: "Rails").call(view_context:)
-      )
+      instance.view_context
     end
   end
 end
 
+# We need to override the randomness of the InlineSvg gem to ensure that the SVG
+# IDs are consistent across test runs and git commits.
 module InlineSvgIdGeneratorRandomnessOverrideForFixtures
   def call
     "non-random-for-fixtures"
